refactor: drop store.ts re-export wall, codemod call sites to slices
Tech-debt R11#3: store.ts had a 26-line back-compat re-export block
forwarding 12 types + 9 helpers from `slices/types` and
`slices/appearanceSlice`. Header comment admitted these were a
pass-through "until we ever flatten further" — flattening now.

Codemod consumers to import directly from the owning slice:
- src/stores/persist.ts: Theme/FontSize/FontSizePx -> slices/types
- src/components/SidebarResizer.tsx: SIDEBAR_WIDTH_*/sanitizeSidebarWidth
  -> slices/appearanceSlice
- src/components/CommandPalette.tsx: resolveEffectiveTheme
  -> slices/appearanceSlice
- src/app-effects/useThemeEffect.ts: resolveEffectiveTheme
  -> slices/appearanceSlice
- tests/appearance.test.ts, tests/store-appearance.test.ts,
  tests/sidebar-resizer-keyboard.test.tsx: same shifts.

Also dropped the unused trailing `export type { State, Actions }`
re-export (no consumers; verified via grep) and pruned the now-unused
appearanceSlice imports inside store.ts (kept only the three helpers
hydrateStore actually calls).

store.ts: 273 -> 225 LoC. typecheck + build clean.
